+++ modules/groups/__init__.py
@@ -8,7 +8,7 @@ def get_priority():
 def get_menu():
     return {
         "display_name": groups_config.module_name, 
-        "name": groups_config.module_name,
+        "module_name": groups_config.module_name,
         "url": "/groups", 
         "external_link": False,
         "priority": groups_config.priority,