+++ modules/search/search.py
@@ -103,5 +103,5 @@ def preserve_current_version():
     """ Preserve current version """
     
     # Check for intermodule dependency
-    if [key['name'] for key in modules.run_ptr if key['name'] == 'versions']:
+    if [key['module_name'] for key in modules.run_ptr if key['module_name'] == 'versions']:
         versions.versions.deploy_current_version()